common: Move RoomNetwork serde in its own file
